requirements.txt
Django>=4.2
django-allauth>=0.58.0
Pillow>=9.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
whitenoise>=6.0
daphne>=3.0.0
